% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elements.R
\name{get_elements}
\alias{get_elements}
\title{Lazily retrieve elements from a Cartesian product}
\usage{
get_elements(l, indices, index_colname = ".element")
}
\arguments{
\item{l}{A named list with the combinations of the Cartesian product.}

\item{indices}{A numeric vector with indices of the elements to retrieve.}

\item{index_colname}{A string with the column name for the indices in the output
data frame. By default, ".element".}
}
\value{
A data frame in which each row is an element from the Cartesian product.
}
\description{
This function lazily retrieves elements from a Cartesian product, given a vector of indices.
}
\examples{
l <- list(color = c("Red", "Blue", "Yellow"),
          shape = c("Square", "Circle"))

get_elements(l, c(2, 4))

}
\references{
Burdsall, T. (2018). \code{lazy-cartesian-product}: .hpp library to efficiently
generate combinations using the Lazy Cartesian Product algorithm. \url{https://github.com/tylerburdsall/lazy-cartesian-product}
}
