% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_cartesian.R
\name{lazy_cartesian}
\alias{lazy_cartesian}
\title{Create a \code{lazy_cartesian} object}
\usage{
lazy_cartesian(l, restrictions = NULL)
}
\arguments{
\item{l}{A named list with the combinations of the Cartesian product.}

\item{restrictions}{A list of lists with restrictions. Defaults to NULL.
See Examples below.}
}
\value{
A \code{lazy_cartesian} object.
}
\description{
This function creates a \code{lazy_cartesian} object, with possible restrictions.
}
\examples{
l <- list(color  = c("Red", "Blue", "Yellow"),
          shape  = c("Square", "Circle"),
          number = 1:3)

r <- list(
  restriction1 = list(color = "Red", shape = "Circle"),
  restriction2 = list(shape = "Square", number = c(1, 3))
)

lc_without_restrictions <- lazy_cartesian(l)
lc_without_restrictions

lc_with_restrictions <- lazy_cartesian(l, r)
lc_with_restrictions

}
\references{
Burdsall, T. (2018). \code{lazy-cartesian-product}: .hpp library to efficiently
generate combinations using the Lazy Cartesian Product algorithm. \url{https://github.com/tylerburdsall/lazy-cartesian-product}
}
