% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{restrict}
\alias{restrict}
\title{Impose a restriction to a Cartesian product}
\usage{
restrict(l, restriction, v_each)
}
\arguments{
\item{l}{A named list with the combinations for the Cartesian product.}

\item{restriction}{A named list that forms a restriction. The name of each
element is a variable. The value of each element is a value or values in the
variable. Multiple elements are joined via `&`.}

\item{v_each}{A vector with the times each variable's values should be repeated. 
See `rlazycartesian:::restrict_multi()`.}
}
\description{
An internal function that imposes a restriction to a Cartesian product
}
\keyword{internal}
