% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{restrict_multi}
\alias{restrict_multi}
\title{Impose multiple restrictions to a Cartesian product}
\usage{
restrict_multi(l, restrictions)
}
\arguments{
\item{l}{A named list with the combinations for the Cartesian product.}

\item{restrictions}{A named list of restrictions.
See `restrict_multi()`.}
}
\description{
An internal function that imposes multiple restrictions to a 
  Cartesian product. See `rlazycartesian:::restrict()`.
}
\keyword{internal}
