% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_size.R
\name{get_size}
\alias{get_size}
\title{Lazily get the total number of elements in a Cartesian product}
\usage{
get_size(l)
}
\arguments{
\item{l}{A named list with the combinations of the Cartesian product.}
}
\value{
A number.
}
\description{
This function lazily retrieves the total number of elements in a Cartesian product
}
\examples{
l <- list(color = c("Red", "Blue", "Yellow"),
          shape = c("Square", "Circle"))

get_size(l)

}
